pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Projeto Integrador - Backend"
authors = [
    {name = "vitormanoelssantana", email = "vitormanoels.santana@gmail.com"},
]
dependencies = ["django>=4.2.21", "djangorestframework>=3.16.0"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
